move resize to tileimage
