+++ src/image.rs
@@ -1,5 +1,5 @@
 use crate::{Config, Error};
-use image::GenericImageView;
+use image::{imageops, GenericImageView};
 use image::{io::Reader as ImageReader, DynamicImage, SubImage};
 use std::ops::RangeInclusive;
 use zorder::coord_of;
@@ -45,6 +45,29 @@ impl<'c> TileImage<'c> {
             tilesize: self.config.tilesize,
             targetrange: targetrangetoslice.clone(),
         }
+    }
+
+    fn _check_dimension(config: &Config, img: &DynamicImage) {
+        if config.endzoomrangetoslice > config.zoomlevel {
+            panic!("Zoom range has value(s) larger than zoom level.");
+        }
+        let (img_width, img_height) = (img.width(), img.height());
+        let max_dimension_size = config.tilesize << config.zoomlevel;
+        if img_width != max_dimension_size || img_height != max_dimension_size {
+            panic!(
+                "Image of size {w}x{h} cannot be split into
+                tiles of size {tile_size} and max zoom level {max_zoom}.",
+                w = img_width,
+                h = img_height,
+                tile_size = config.tilesize,
+                max_zoom = config.zoomlevel,
+            );
+        }
+    }
+    
+    pub fn resize(&self, img: &DynamicImage, width: u32, height: u32) -> DynamicImage {
+        Self::_check_dimension(self.config, img);
+        img.resize(width, height, imageops::FilterType::Lanczos3)
     }
 }
